ノート編集ページのボタンとサイドバーのリファクタリング
