+++ next/src/pages/dashboard/notes/[slug]/edit.tsx
@@ -1,33 +1,16 @@
-import CloseIcon from '@mui/icons-material/Close'
-import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline'
-import KeyboardDoubleArrowRightIcon from '@mui/icons-material/KeyboardDoubleArrowRight'
-import PlayArrowOutlinedIcon from '@mui/icons-material/PlayArrowOutlined'
-import SellOutlinedIcon from '@mui/icons-material/SellOutlined'
-import {
-  AppBar,
-  Autocomplete,
-  Box,
-  Chip,
-  Divider,
-  Drawer,
-  Fade,
-  IconButton,
-  Stack,
-  TextField,
-  Tooltip,
-  Typography,
-} from '@mui/material'
+import { AppBar, Box, Fade, Stack, TextField, Typography } from '@mui/material'
 import axios from 'axios'
 import type { NextPage } from 'next'
 import { useRouter } from 'next/router'
 import { useCallback, useEffect, useMemo, useState, useRef } from 'react'
 import { Helmet, HelmetProvider } from 'react-helmet-async'
 import { useForm, SubmitHandler, Controller } from 'react-hook-form'
-import ConfirmDialog from '@/components/common/ConfirmDialog'
 import Error from '@/components/common/Error'
 import Loading from '@/components/common/Loading'
 import MarkdownText from '@/components/note/MarkdownText'
 import { MarkdownToolbar } from '@/components/note/MarkdownToolbar'
+import NoteEditorButtons from '@/components/note/NoteEditorButtons'
+import NoteEditorSidebar from '@/components/note/NoteEditorSidebar'
 import { NoteEditorToolbar } from '@/components/note/NoteEditorToolbar'
 import { RestoreConfirmationBanner } from '@/components/note/RestoreConfirmagionBanner'
 import { useAuthContext } from '@/hooks/useAuthContext'
@@ -36,11 +19,10 @@ import useEnsureAuth from '@/hooks/useEnsureAuth'
 import { useLocalStorage } from '@/hooks/useLocalStorage'
 import { useMyNote } from '@/hooks/useMyNote'
 import { useSnackbarState } from '@/hooks/useSnackbarState'
-import { useTags } from '@/hooks/useTags'
 import { styles } from '@/styles'
 import { handleError } from '@/utils/handleError'
 
-interface NoteFormData {
+export interface NoteFormData {
   title: string
   description: string
   content: string
@@ -58,27 +40,16 @@ const EditNote: NextPage = () => {
   const noteSlug = typeof slug === 'string' ? slug : undefined
   const { noteData, noteError } = useMyNote()
 
-  const { tagsData } = useTags()
-
   const [isPreviewActive, setIsPreviewActive] = useState<boolean>(false)
   const [statusChecked, setStatusChecked] = useState<boolean>(false)
   const [isFetched, setIsFetched] = useState<boolean>(false)
   const [isLoading, setIsLoading] = useState<boolean>(false)
-  const [maxLengthError, setMaxLengthError] = useState<boolean>(false)
-  const [maxTagsError, setMaxTagsError] = useState<boolean>(false)
-  const [formatError, setFormatError] = useState<boolean>(false)
-  const [charCount, setCharCount] = useState<number>(0)
   const [openSidebar, setOpenSidebar] = useState<boolean>(false)
   const [content, setContent] = useState<string>('')
   const [inputTags, setInputTags] = useState<string[]>([])
-  const [inputValue, setInputValue] = useState<string>('')
 
   const { getElapsedSeconds } = useDuration()
   const [previousSeconds, setPreviousSeconds] = useState<number>(0)
-
-  const [openDeleteConfirmDialog, setOpenDeleteConfirmDialog] =
-    useState<boolean>(false)
-  const [noteSlugToDelete, setNoteSlugToDelete] = useState<string | null>(null)
 
   const textareaRef = useRef<HTMLTextAreaElement>(null)
 
@@ -99,38 +70,6 @@ const EditNote: NextPage = () => {
     const newValue = e.target.value
     setContent(newValue)
     if (!restoreContent) saveContent(newValue)
-  }
-
-  const handleInputChange = (newInputValue: string) => {
-    setFormatError(false)
-    setMaxLengthError(false)
-    setMaxTagsError(false)
-
-    if (!newInputValue) {
-      setInputValue('')
-      return
-    }
-
-    const formatRegex = /^[a-zA-Z0-9ａ-ｚＡ-Ｚ０-９ぁ-んァ-ン一-龯]+$/
-    if (!formatRegex.test(newInputValue)) {
-      setFormatError(true)
-      return
-    }
-
-    const maxTagLength = 20
-    if (newInputValue.length > maxTagLength) {
-      setMaxLengthError(true)
-      return
-    }
-
-    const maxTags = 5
-    if (inputTags.length >= maxTags) {
-      setMaxTagsError(true)
-      return
-    }
-
-    setInputValue(newInputValue)
-    setIsChanged(true)
   }
 
   const [imageSignedIds, setImageSignedIds] = useState<
@@ -263,38 +202,6 @@ const EditNote: NextPage = () => {
     } finally {
       setIsLoading(false)
     }
-  }
-
-  const handleDeleteNote = (noteSlug?: string) => {
-    if (!noteSlug) return
-
-    setNoteSlugToDelete(noteSlug)
-    setOpenDeleteConfirmDialog(true)
-  }
-
-  const handleDeleteConfirm = async () => {
-    if (!noteSlugToDelete) return
-
-    const url = `${process.env.NEXT_PUBLIC_API_BASE_URL}/my_notes/${noteSlug}`
-    const headers = { Authorization: `Bearer ${idToken}` }
-
-    try {
-      await axios.delete(url, { headers })
-      router.push('/dashboard')
-    } catch (err) {
-      const { errorMessage } = handleError(err)
-      setSnackbar({
-        message: errorMessage,
-        severity: 'error',
-        pathname: router.pathname,
-      })
-    } finally {
-      setOpenDeleteConfirmDialog(false)
-    }
-  }
-
-  const handleCloseDeleteConfirmDialog = () => {
-    setOpenDeleteConfirmDialog(false)
   }
 
   if (noteError) {
@@ -631,304 +538,26 @@ const EditNote: NextPage = () => {
                 />
 
                 {/* ボタン（プレビュー表示・タグ入力欄表示・画像追加・削除） */}
-                <Box
-                  sx={{
-                    display: 'flex',
-                    justifyContent: 'space-around',
-                    position: 'fixed',
-                    width: '300px',
-                    bottom: '30px',
-                    transition: '0.2s',
-                    zIndex: 1000,
-                  }}
-                >
-                  <Tooltip
-                    title={
-                      !isPreviewActive ? 'プレビューを表示' : 'エディタのみ表示'
-                    }
-                  >
-                    <IconButton
-                      onClick={togglePreviewDisplay}
-                      sx={{
-                        backgroundColor: !isPreviewActive
-                          ? 'white'
-                          : 'primary.main',
-                        boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)',
-                        '&:hover': {
-                          backgroundColor: 'backgroundColor.hover',
-                        },
-                      }}
-                    >
-                      <PlayArrowOutlinedIcon
-                        sx={{
-                          fontSize: 30,
-                          color: !isPreviewActive ? undefined : 'white',
-                        }}
-                      />
-                    </IconButton>
-                  </Tooltip>
-                  <Tooltip title="タグ・概要を登録">
-                    <IconButton
-                      onClick={toggleSidebar}
-                      sx={{
-                        backgroundColor: 'white',
-                        boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)',
-                        width: '46px',
-                        height: '46px',
-                        '&:hover': {
-                          backgroundColor: 'backgroundColor.hover',
-                        },
-                      }}
-                    >
-                      <SellOutlinedIcon />
-                    </IconButton>
-                  </Tooltip>
-                  <Tooltip title="ノートを削除">
-                    <IconButton
-                      onClick={() => handleDeleteNote(noteSlug)}
-                      sx={{
-                        backgroundColor: 'white',
-                        boxShadow: '0 4px 8px rgba(0, 0, 0, 0.2)',
-                        width: '46px',
-                        height: '46px',
-                        '&:hover': {
-                          backgroundColor: 'backgroundColor.hover',
-                        },
-                      }}
-                    >
-                      <DeleteOutlineIcon sx={{ color: '#f28b82' }} />
-                    </IconButton>
-                  </Tooltip>
-                </Box>
+                <NoteEditorButtons
+                  isPreviewActive={isPreviewActive}
+                  togglePreviewDisplay={togglePreviewDisplay}
+                  toggleSidebar={toggleSidebar}
+                />
               </Box>
             </Stack>
           </Box>
         </Fade>
 
-        {/* ノート削除の確認画面 */}
-        <ConfirmDialog
-          open={openDeleteConfirmDialog}
-          onClose={handleCloseDeleteConfirmDialog}
-          onConfirm={handleDeleteConfirm}
-          message={'ノートを削除しますか？'}
-          confirmText="実行"
-        />
-
-        <Drawer
+        {/* サイドバー（タグ・概要入力フィールド） */}
+        <NoteEditorSidebar
           open={openSidebar}
-          anchor="right"
-          variant="persistent"
-          sx={{
-            borderTop: 'none',
-            flexShrink: 0,
-            '& .MuiDrawer-paper': {
-              width: { xs: '100%', md: '400px' },
-              boxSizing: 'border-box',
-            },
-          }}
-        >
-          <Box
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              height: '64px',
-            }}
-          >
-            <IconButton onClick={toggleSidebar} sx={{ ml: 2 }}>
-              <KeyboardDoubleArrowRightIcon />
-            </IconButton>
-            <Typography
-              color={'text.light'}
-              fontWeight={'bold'}
-              fontSize={18}
-              sx={{ ml: 1 }}
-            >
-              タグ・概要の登録
-            </Typography>
-          </Box>
-          <Divider />
-
-          <Box
-            css={styles.pageMinHeight}
-            style={{ width: '100%', padding: 20 }}
-          >
-            <Stack spacing={3}>
-              {/* タグ入力フィールド */}
-              <Box>
-                <Typography
-                  sx={{ fontWeight: 'bold', color: 'text.light', p: 1 }}
-                >
-                  タグ
-                </Typography>
-                <Box
-                  sx={{
-                    border: '1px solid',
-                    borderRadius: 2,
-                    borderColor: 'divider',
-                    p: 2,
-                  }}
-                >
-                  <Controller
-                    name="tags"
-                    control={control}
-                    render={({ field }) => (
-                      <Autocomplete
-                        {...field}
-                        multiple
-                        freeSolo
-                        options={tagsData?.map((tag) => tag.name)}
-                        value={inputTags}
-                        onChange={(_, newTagNames) => {
-                          setFormatError(false)
-                          setMaxLengthError(false)
-                          if (maxTagsError) {
-                            setMaxTagsError(false)
-                            return
-                          }
-
-                          if (newTagNames.length > 5) {
-                            setMaxTagsError(true)
-                            return
-                          }
-
-                          setInputTags(newTagNames)
-                        }}
-                        renderTags={(value, getTagProps) => (
-                          <Box>
-                            {value.map((option, index) => {
-                              const { key, ...tagProps } = getTagProps({
-                                index,
-                              })
-                              return (
-                                <Chip
-                                  label={option}
-                                  variant="outlined"
-                                  deleteIcon={
-                                    <CloseIcon
-                                      sx={{ width: '16px', height: '16px' }}
-                                    />
-                                  }
-                                  sx={{ fontSize: '12px', mr: 1 }}
-                                  key={key}
-                                  {...tagProps}
-                                />
-                              )
-                            })}
-                          </Box>
-                        )}
-                        renderOption={(props, option) => {
-                          const tag = tagsData.find(
-                            (tag) => tag.name === option,
-                          )
-
-                          return (
-                            <li {...props}>
-                              {option}{' '}
-                              {tag?.notesCount ? `(${tag.notesCount})` : ''}
-                            </li>
-                          )
-                        }}
-                        inputValue={inputValue}
-                        onInputChange={(_, value) => handleInputChange(value)}
-                        renderInput={(params) => (
-                          <TextField
-                            {...params}
-                            placeholder={
-                              !inputTags.length
-                                ? 'タグは5つまで登録できます'
-                                : ''
-                            }
-                            error={
-                              maxTagsError || formatError || maxLengthError
-                            }
-                            helperText={
-                              (maxTagsError &&
-                                'タグは最大5つまで登録できます') ||
-                              (formatError &&
-                                'タグに記号とスペースは使用できません') ||
-                              (maxLengthError && 'タグの最大文字数は20文字です')
-                            }
-                            sx={{
-                              '& .MuiInputBase-root': {
-                                p: 0,
-                              },
-                              '& .MuiOutlinedInput-input': {
-                                p: 0,
-                                minWidth: '200px',
-                              },
-                              '& .MuiOutlinedInput-notchedOutline': {
-                                border: 'none',
-                              },
-                              '& .MuiFormHelperText-root': {
-                                position: 'absolute',
-                                top: '-52px',
-                                left: '36px',
-                                color: 'error',
-                              },
-                            }}
-                          />
-                        )}
-                      />
-                    )}
-                  />
-                </Box>
-              </Box>
-
-              {/* 概要入力フィールド */}
-              <Box>
-                <Typography
-                  sx={{ fontWeight: 'bold', color: 'text.light', p: 1 }}
-                >
-                  概要
-                </Typography>
-                <Box
-                  sx={{
-                    border: '1px solid',
-                    borderRadius: 2,
-                    borderColor: 'divider',
-                    p: 1,
-                  }}
-                >
-                  <Controller
-                    name="description"
-                    control={control}
-                    rules={validationRules.description}
-                    render={({ field, fieldState }) => (
-                      <>
-                        <TextField
-                          {...field}
-                          type="textarea"
-                          error={fieldState.invalid}
-                          helperText={fieldState.error?.message}
-                          placeholder="ノートの概要を入力してください"
-                          multiline
-                          fullWidth
-                          rows={12}
-                          onChange={(e) => {
-                            field.onChange(e)
-                            setCharCount(e.target.value.length)
-                          }}
-                          sx={{
-                            '& .MuiOutlinedInput-notchedOutline': {
-                              border: 'none',
-                            },
-                          }}
-                        />
-                        <Typography
-                          variant="caption"
-                          color={charCount > 200 ? 'error' : 'text.light'}
-                          sx={{ textAlign: 'right', display: 'block' }}
-                        >
-                          {charCount} / 200文字
-                        </Typography>
-                      </>
-                    )}
-                  />
-                </Box>
-              </Box>
-            </Stack>
-          </Box>
-        </Drawer>
+          onClose={toggleSidebar}
+          control={control}
+          inputTags={inputTags}
+          setInputTags={setInputTags}
+          setIsChanged={setIsChanged}
+          descriptionValidationRule={validationRules.description}
+        />
       </Box>
     </>
   )
